Only Watch Proxy when running on OpenShift
HIVE-2198 / #1999 / 253408ed introduced a `Watch()` on `Proxy` so we can
update environment variables and CA bundles when it changes. But `Proxy`
is an OpenShift CRD, so that `Watch()` blows up the operator when
running on vanilla k8s (and similar).

This commit conditions that `Watch()` such that we only add it if we're
running on OpenShift.

HIVE-2210

(cherry picked from commit f1f63cbfad47b78d69f9a8f98aa481ae0a2053aa)