Dummy ca-bundle.crt for non-OpenShift
If we're not running on OpenShift, the magic label doesn't cause
OpenShift to populate the merged CA bundle ConfigMap. This would cause
the ConfigMap mount to fail, and provisions wouldn't start. This commit
defaults that key to empty so the file will always exist and the mount
will work. (Note that `update-ca-trust` properly ignores an empty file.)

HIVE-2210