Add timestamps to messages in e2e tests
Debugging e2e failures -- especially those with a timing component --
will be easier if we have date stamps on our output messages. Make it
so.

Note that this commit is only overriding the `echo` command. Output from
subcommands won't be affected. That may turn out to be a problem, but
it's one we can solve later.