Comment: freeze ClusterAutoscaler configuration by MachinePool controller
The MachinePool controller should not be in the business of configuring
the (singleton) ClusterAutoscaler. We've got some minimal setup in the
code that's considered necessary if the (any) MachinePool has
Autoscaling enabled; but further configuration should be left to the
user -- whether in-cluster or from the hub via syncsets.

HIVE-2220