MachinePool: Allow Platform edits via override
Normally we forbid updates to MachinePool.Spec.Platform via an admission
hook because, while we would roll out the changes to the MachineSets on
the target cluster, MCO ignores those changes.

With this commit, you can add an annotation to a MachinePool
```
hive.openshift.io/override-machinepool-platform: "true"
```
to bypass this check. By doing so, you are taking responsibility for
forcing rollout of any changes, e.g. by deleting the Machines on the
target cluster.

HIVE-2024