e2e: Save interim manifests
It was annoying to not be able to get things like clusterdeployment
manifests for e2e runs that completed successfully. Now e2e will save
those manifests twice: once right after the ClusterDeployment has been
created and once on cleanup. (As today, the latter will only be
interesting if the test fails.)