Skip hive-trusted-cabundle ConfigMap creation on deleted ns
Deleting a namespace containing a ClusterDeployment and its associated
artifacts results in an impossible situation: the deletion of the
ClusterDeployment wants to trigger a deprovision, but we cannot create
the necessary artifacts (e.g. ClusterDeprovision CR, deprovision Pod) in
a namespace that is terminating. We have logic to skip deprovisioning
and simply leak the cloud resources in this case.

...but we forgot to incorporate this logic for the hive-trusted-cabundle
ConfigMap (introduced by 253408ed8 / #1999), and would hot-loop
attempting and failing to create it, never reaching the aforementioned
"give up" logic, resulting in wedging the namespace. This commit adds
the logic, resulting in the same behavior as before.

NOTE: It is still not supported / a good idea to delete a namespace
containing extant ClusterDeployments, as this leaks cloud resources as
described above.

HIVE-2249