MachinePool: hack name of vSphere default worker MachineSet
With recent installer changes to accommodate zones in vSphere, the name
of the default worker pool in OCP 4.13+ is now suffixed with a `-0` even
when only one zone is in play. Hive's day 2 MachinePool logic needs to
be able to generate the correct MachineSet name when talking to a 4.13+
cluster -- but also when talking to a 4.12- cluster. This commit adds in
a special case to discover and account for this difference where it
exists.

HIVE-2254