MachinePool: Properly handle labels & taints
With this commit, MachinePool grows two new statusfields: `OwnedLabels`
and `OwnedTaints`. These keep track of the labels and taints that were
added to the remote MachineSets at the behest of this MachinePool.
Conversely, this allows us to identify which ones were added on the
remote side, and thus shouldn't be touched by our controller.

We also fix the logic for matching taints, which are "keyed" by the
combination of the `Key` and `Effect` fields.

To summarize the behavior with this commit in play:
- Labels and taints listed in MachinePool.Spec are synced to the remote
MachineSets.
- Labels and taints that are *not* in MachinePool.Spec are removed from
the remote MachineSets *iff* they are present in
MachinePool.Status.Owned{Labels|Taints}.
- Upon successful syncing, we update
MachinePool.Status.Owned{Labels|Taints} to match the spec.

The effect is that editing MachinePool.Spec to remove a label or taint
should cause that label/taint to be removed from the remote MachineSets,
BUT labels/taints that only exist on the remote MachineSets remain
untouched.

HIVE-2035
HIVE-2276
HIVE-2279