e2e (AWS): Allow env var to override creds file
Resolve a long-standing TODO which allows AWS environment variables
`AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY` to be used when present,
even if a credentials file exists, as the gods intended.

hiveutil already uses the appropriate logic such that, if a creds file
is not supplied, it looks for the env vars. We just needed to fix it so
we don't pass that arg in at all if the env vars are set.

(This was an incidental requirement of HIVE-2234: the AWS credentials
from the hypershift installation were taking precedence for managed DNS,
which then bounced because the expected hosted zone is in a different
account.)