e2e-pool: Start real pool provisions after inventory added
Goal: reduce e2e-pool wallclock time by ~35m.

Problem Statement: When ClusterPool inventory
(ClusterDeploymentCustomization) testing was added to e2e-pool (4fddbe7
/ #1672), it triggered ClusterPool's staleness algorithm such that we
were actually wasting a whole cluster while waiting for the real pool to
become ready. Grab a cup of coffee...

To make the flow of the test a little bit easier, we were creating the
real pool, then using its definition to generate the fake pool
definition -- which does not have inventory -- and then adding inventory
to the real pool.

But if you add or change a pool's inventory, we mark all its clusters
stale. So because of the flow above, when we initially created the real
pool without inventory, it started provisioning a cluster. Then when we
updated it (mere seconds later, if that), that cluster immediately
became stale.

Now, the way we decided to architect replacement of stale clusters, we
prioritize _having claimable clusters_ over _all clusters being
current_. Thus in this scenario we were actually ending up waiting until
the stale cluster was fully provisioned before deleting it and starting
over with the (inventory-affected) cluster.

Solution: Create the real pool with an initial `size=0`. Scale it up to
`size=1` _after_ adding the inventory.