e2e (AWS): Default BASE_DOMAIN for new CI account
Our e2e suite for AWS was setting a default base domain that (AFAIK) we
don't have control over, so we were overriding it in the test config.
We've recently switched to a new AWS account for these tests. Change the
default base domain accordingly so we don't need to override it in the
test config.

HIVE-2259