e2e-pool: Plumb in BASE_DOMAIN
hiveutil still defaults the base domain to new-installer.openshift.com,
which is probably what we want since we (hive engineers) are the ones
invoking it by hand, using the hive-team cluster as a hub, and will thus
prefer less typing.

But now that we're switching to a new AWS account for CI, we want the
AWS-based e2e suites to use the new base domain. We recognize and
default a `$BASE_DOMAIN` variable for e2e itself, but weren't yet
plumbing it through to e2e-pool. Fix.

HIVE-2259