e2e-pool: Improve cleanup artifact gathering
We've been seeing some weird timeouts in e2e-pool where the test itself
succeeds, but clusterclaim deletion hangs until the job times out and is
killed by the monitor. Previously we had no good way to debug this
because we had last captured manifests & logs *before* issuing the
clusterclaim deletion.

With this commit, we:
- Do that deletion (and that of the clusterpools) in the background.
- Re-capture logs and manifests periodically while we're waiting for
cleanup to complete.