Collapse duplicate taints on MachineSets
Since taints are provided as lists, duplicate entries are possible -
which can cause confusion during sync. While we will not touch the
MachinePool Spec, on both the remote MachineSet and generated
MachineSet, we will collapse the duplicate entries.
Unique Identifier for Taints is Key+Effect. If there are duplicates with
different Value of a taint entry, the Value from the first entry
encountered will be preserved, and the rest will be collapsed.

This commit also includes test cases that cover duplicates both on
MachinePool as well as remote MachineSet, w.r.t. updating as well as
deleting.

Sorting of labels and taints is removed from this commit since it was
not deemed necessary here.