Add configmap for Hive SLO grafana dashboard
We would like to add Hive SLO grafana dashboard to Prod App-SRE Grafana.
This will be done by injecting the JSON for the said dashboard to a
ConfigMap and point the saas file for Hive to manage the
resourceTemplate for this ConfigMap.

For future changes to this dashboard, it should be easy enough to
generate a new JSON and replace it in our ConfigMap, retaining the
metadata.