AWS deprovision: Plumb HostedZoneRole for shared VPC
The shared VPC flow added the need for the installer to use a
HostedZoneRole to manage certain DNS records. This is provided at
installation time via the install-config, and is included in the
metadata.json produced by the installer when provisioning completes.

The uninstall process requires this HostedZoneRole during deprovisioning
to perform the corresponding cleanup -- without it, those records are
leaked. So we need to plumb that datum through from the metadata.json to
the uninstall process.

To make this more general-purpose for adoption and disaster recovery,
and for retrofitting existing clusters that were created before this
commit, we also attempt to pull the HostedZoneRole out of the
install-config if it is unset.

NOTE: We will refuse to proceed with deprovision if this field is not
set. Once this change lands, we should audit existing clusters to make
sure they all get retrofitted; and manually retrofit any others (with
the empty string).

HIVE-2297