Initialize clusterMetadata as late as possible
When retrofitting cd.spec.clusterMetadata.platform.aws.hostedZoneRole
(see #2100), we need to initialize the parent structures only once we
know we're going to be setting the hostedZoneRole field. Otherwise, if
parsing the install-config fails for some reason, we may end up trying
to Update() the cd with that `aws` section empty, which is forbidden by
webhook.

Turns out we also made this mistake when setting the azure
resourceGroupName as well (see #1997). Fix.

HIVE-2297