AWS deprovision: Plumb ClusterDomain
ClusterUninstaller.ClusterDomain has existed for a long time, but was
never used before the advent of shared VPC. In this new flow, if it is
absent, DNS entries in the private hosted zone are not cleaned up in the
destroy flow. This commit adds a new CLI parameter, `--cluster-domain`,
to `hiveutil aws-tag-deprovision`. It is plumbed through by adding
`spec.BaseDomain` to the ClusterDeprovision CRD and constructing the
value by joining that to the `spec.ClusterName`.

TODO: Our platform-specific ClusterDeprovision sub-structs for Alibaba
and IBMCloud each have a `BaseDomain` field. We should transition those
to using this new platform-agnostic field -- but later. See HIVE-2298.

HIVE-2297