Add reachable check to (remote client) Builder
With the controller-runtime upgrade from 8cfdcca, our remote client
`Builder` interface's `Build()` method got lazier. Before, if the remote
client was unreachable, `Build()` itself would fail. Now it won't fail
until the resulting client is actually used. This resulted in our
"unreachable" controller incorrectly marking the ClusterDeployment's
APIURLOverride as reachable when it wasn't, which had the ripple effect
of making the clustersync controller fail early by trying to use that
URL rather than falling back to the default.

This was exposed in situations where the DNS for the APIURLOverride was
being configured via a syncset: the clustersync controller tried to use
a URL that wouldn't resolve, so it couldn't sync the resource that would
make that DNS work.

To fix, this commit adds a check in the `Build()` method of the `Builder`
interface. This internally builds the client and then uses it to query
the `/apis` of the remote client.

HIVE-2300

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>