e2e: syncset test: retry on conflict
We've recently seen more than one occurrence of 409 conflict flakes in
this test suite, where a previously-retrieved instance of
ClusterDeployment had a stale resourceVersion by the time it was
Update()d back to the server.

This commit adds a RetryOnConflict wrapper around both Update()s in this
test, which should hopefully reduce errors to "real" ones.

HIVE-2294