Fix fake cluster version
Fake clusters report their version based on a dummy CVO incorporated
into their special remote client builder. Previously we were hardcoding
that value to "4.6.8" in that builder. With this commit, we plumb in the
value from cd.status.installVersion -- which gets set by the
imageset/installer process even for fake clusters. So now if you watch
the `VERSION` field for a fake cluster, it behaves the same as for a
real cluster: empty until the cluster is `Provisioned` and the
clusterversion controller kicks in again, whereupon it is set to the
correct value based on the input release image.

HIVE-2312