Refer to logs for AWS auth errors
Today when an AWS API call fails during hibernation, the "Hibernating"
ClusterDeployment status condition is updated with the error message.
Sometimes this message is sent through a scrubber to ensure it does not
cause thrashing if it happens to be unique every time (e.g. if it
includes a timestamp or transaction ID). One example is an authorization
failure, e.g. when the IAM user lacks a necessary role to perform the
API action. In such cases, the encoded failure message is redacted for
purposes of the status condition. The full (encoded) message is in the
logs... but the user doesn't necessarily know that. So with this commit,
we tell them by adding a little note to the status condition message.

HIVE-2319