DNM: Test e2e-gcp required-ness
We're trying to make e2e-gcp (and -azure) mandatory iff they are
triggered: https://github.com/openshift/release/pull/44171

This commit does two things:
- No-op change to a GCP-specific source file to cause e2e-gcp to be
triggered.
- Change to postinstall that should make e2e-gcp fail

The idea will be to `lgtm` this PR to demonstrate that it doesn't merge!

(Then we'll want to watch a normal PR that we _do_ want to merge and
make sure that not running e2e-gcp at all doesn't block it.)