fix `make verify[-codegen]`
First:

`verify-codegen.sh` was a thin wrapper around `update-codegen.sh` that
passed a `--verify-only` flag to the generators. But generators wasn't
the only thing in `update-codegen.sh`, which led to inconsistencies.
This commit changes the strategy entirely:

We're getting rid of the `verify-codegen.sh` wrapper. Instead
`verify-codegen` will run `update-codegen` and then check if anything
changed. (This is similar to how `verify-vendor` works.)

Second:

Apparently the generators invoked by `update-codegen.sh` used to put the
generated files under vendor/, so we had an extra line to copy them back
to the non-vendor/ side. At some point this seems to have changed such
that the generators put the files in non-vendor/. So that copy was
actually _undoing_ what the generators had just done.

We're getting rid of that copy-back entirely. So now, `make
update-codegen` shouldn't touch the vendor/ directory; and `verify` will
catch discrepancies here (i.e. you forgot to run `make vendor`) via
`verify-vendor` rather than in `verify-codegen` itself (which wasn't
working anyway).

Third:

`make update-codegen` wouldn't fail if something in it went wrong. Add
`-e` to `update-codegen.sh` so it does. (This is _almost_ totally
unrelated to the rest of what's going on here.)