Annotation to hold uninstall pod
To destroy a cluster, you delete the ClusterDeployment. Hive responds as
follows:
- Create a ClusterDeprovision
- Create a Job
- ...which creates a Pod
- ...which has a Container
- ...which runs the installer's destroy code

Under normal circumstances, when the destroy process finishes, the Pod
gets garbage collected very quickly -- often before logs can even finish
streaming, if they're being watched. This can make debugging uninstall
issues difficult.

With this commit, we add support for an annotation on ClusterDeployment:

`hive.openshift.io/hold-uninstall-pod: true`

This annotation is copied to the ClusterDeprovision, and when hive sees
it there, it will add that key (`hive.openshift.io/hold-uninstall-pod`)
as a finalizer to the uninstall pod. This will cause it to remain in
`Terminating` state until the finalizer is removed, *which must be done
"manually" by the consumer*.

HIVE-2309