GCP deprovision: Plumb NetworkProjectID for shared VPC
The shared VPC flow added the need for the installer to use a
NetworkProjectID to manage certain resources. This is provided at
installation time via the install-config, and is included in the
metadata.json produced by the installer when provisioning completes.

The uninstall process requires this NetworkProjectID during
deprovisioning to perform the corresponding cleanup -- without it, those
resources are leaked. So we need to plumb that datum through from the
metadata.json to the uninstall process.

To make this more general-purpose for adoption and disaster recovery,
and for retrofitting existing clusters that were created before this
commit, we also attempt to pull the NetworkProjectID out of the
install-config if it is unset.

NOTE: We will refuse to proceed with deprovision if this field is not
set. Once this change lands, we should audit existing clusters to make
sure they all get retrofitted; and manually retrofit any others (with
the empty string).

HIVE-2334