Minimal install mode (don't pull release or oc images)
Certain environments, like ARO:
- Always use InstallerImageOverride
- Don't care to must-gather spokes on install failures

The latter is the only thing we use the `oc` command for.
And the release image is only used to discover the URI for the installer
image and the `oc` image.

To reduce burdens like ensuring those images are available and secure,
this commit introduces a (back door, not-officially-supported) "minimal
install mode". Activate it by doing two things:

1. Set ClusterDeployment annotation
`hive.openshift.io/minimal-install-mode: True`
2. Set `ClusterDeployment.Spec.Provisioning.InstallerImageOverride`

This will cause hive to wend its way through the provisioning flow
without pulling the release or `oc` images. It will use the
InstallerImageOverride as the source of the `openshift-install` binary.

NOTE: This introduces a few limitations/oddities:
- You cannot use release image verification. (There's no release image
to verify.)
- Failed installs will not attempt to must-gather the spoke, even if you
have the rest of the install log capture stuff configured.
- `ClusterDeployment.Status.CLIImage` will be set to `"<NONE>"`.
- `ClusterDeployment.Status.InstallVersion` will not be set.
Consequently...
- The `cluster_version` label on the
`hive_cluster_deployment_install_failure_total` and
`hive_cluster_deployment_install_success_total` metrics will be
`"unspecified"` for these clusters.

HIVE-2208