Annotation to override installer image name
We find the image we use to obtain the `openshift-install` binary based
on its name in release metadata.

```
        {
          "name": "installer",   <== THIS NAME
          "annotations": {
            "io.openshift.build.commit.id": "ead355d7836a8a24613d062bf06ab0a6eef2765c",
            "io.openshift.build.commit.ref": "",
            "io.openshift.build.source-location": "https://github.com/openshift/installer"
          },
          "from": {
            "kind": "DockerImage",
            "name": "quay.io/openshift-release-dev/ocp-v4.0-art-dev@sha256:b2942182a31a0591dcbf3a3eec10c363d9d1f05dd75b98e1f8ac7f486b0ce25e"
          },
          "generation": null,
          "importPolicy": {},
          "referencePolicy": {
            "type": ""
          }
        },
```
By default, we use the image identified as `"installer"`; or
`"baremetal-installer"` for bare metal installations.

Alternative or experimental installer images may be included in certain
release images alongside these well-known ones. This commit provides a
mechanism for consuming such an image instead of the default: in your
ClusterDeployment, set the annotation
`hive.openshift.io/installer-image-name-override` to the release
metadata name of the alternative image you wish to use.

This is currently for internal use only, and will not be documented or
officially supported.

HIVE-2342