Hack/fix for code generator bugs/changes
For some reason in this update of code-generator the shell scripts are
not executable. Make it so. However, we have to revert once the
generators finish or `verify` will complain.

Also this version of codegen removes existing deepcopy files under the
vendor/ directory. This probably isn't a bad thing, but it does force us
to be more careful about the ordering of our rules for `verify`: we have
to `verify-vendor` after we `verify-codegen` to restore those files
before we `verify-imports`, which entails `build`, which fails if those
deepcopy files are absent.

HIVE-2324