Remove SyncSetInstance cleanup code
Remove the code to clean up legacy SyncSetInstance CRs and CRD, which
were supplanted by ClusterSync via #1093 ~3y ago.

The consequence is that, if a customer upgrades from a pre-ClusterSync
version of hive to a version at/after this commit, the SyncSetInstance
CRs and CRD will be orphaned, unnecessarily cluttering etcd.

It has been long enough that we consider the probability of this to be
quite low. If it does happen, there is a simple workaround: manually
delete the CRs and CRD.

Why do this at all? I'm glad you asked. Since the SyncSetInstance CRD
existed in the now-deprecated v1beta1 apiVersion of
apiextensions.k8s.io, our API call to discover it was showing up on
customers' radar and making them uncomfortable. At this point the risk
of orphaning resources is the lesser evil.

HIVE-2025