Read-only root file system
Add `readOnlyRootFilesystem` to our Deployments and StatefulSets. Why?
Cause it's more securer.

One thing I needed to do to make this work was to add a volume for
`/tmp`, which I guess is by default part of `/`. This resulted in an
actual problem in hive-controllers creating a Session for AWS... but I
did it for all the things just in case.

This scares me a little bit. For one thing, any time we mess with
`securityContext`s we seem to break our friends running hive under
vanilla k8s. But also, who knows what corner cases are expecting to be
able to write somewhere in the root file system? We won't find out if,
like, trying to use additional certs under OpenStack is a problem until
some customer tries it.

HIVE-2350