Add Fixed Labels to metrics doc
It can be confusing for consumers wanting to add labels to metrics with
optional label support without being aware of the fixed labels already
reported by those metrics, this commit documents the same to resolve the
problem.

There's also the fix for cluster sync failing metric - recently it
allows additional labels but document update was pending.