SyncSet: Parameterized CD labels in resources
Via a new knob, `[Selector]SyncSet.Spec.EnableResourceTemplates: True`
(that's a bool), it is now possible to use golang
[text/template](https://pkg.go.dev/text/template)-isms in
`[Selector]SyncSet.Spec.Resources[]` string values (not keys! do you
need that?). Note, however, that there is no "dot" (data object) so the
out-of-the-box functionality won't convey a lot of power.

However, this commit introduces one custom function: `fromCDLabel`. If
you include a string like

`{{ fromCDLabel "any.clusterdeployment/label-key" }}`

in a resource, hive will substitute it with the value of that label from
the ClusterDeployment owning the [Selector]SyncSet being processed. For
example:

```yaml
---
apiVersion: hive.openshift.io/v1
kind: SyncSet
metadata:
  name: mygroup
spec:
  clusterDeploymentRefs:
  - name: ClusterName

  resourceApplyMode: Upsert

  enableResourceTemplates: true

  resources:
  - apiVersion: v1
    kind: ConfigMap
    metadata:
      name: mycm
      namespace: myns
    data:
      platform: the platform is {{ fromCDLabel "hive.openshift.io/cluster-platform" }}
```

...might result in the following ConfigMap on the spoke cluster:

```yaml
apiVersion: v1
kind: ConfigMap
metadata:
  name: mycm
  namespace: myns
data:
  platform: the platform is gcp
```

If the CD has no labels, or if there is no label with the specified key,
the empty string is substituted.

Errors parsing or processing the template will cause the SyncSet to fail
and will be bubbled up in the ClusterSync status as usual.

HIVE-1748