Unstructured OpenstackProviderSpec
Another attempt to address the move of OpenstackProviderSpec from
cluster-api-provider-openstack to openstack/api/machine. See #2117 for
background. In that PR, we simply replaced our imports. The result is
that we could only decode the _new_ thing. The problem is that we may
still get the _old_ thing if we're talking to an old cluster.

We only need the decoding to steal the osImage from the master machines
so we can use that value for new workers. Rather than importing both
schemata, trying to figure out which version we're talking to, and
decoding against the correct one, this commit unmarshals the
providerSpec as raw JSON into an Unstructured object and explicitly
paths into it map-wise.

On the creation side, we're told that the spoke cluster itself will
understand the new thing even if it's an old version (MAPI also
unmarshals the JSON raw). So it should be fine to continue using recent
vendored installer code to generate MachineSets (which will contain
new-apiVersion OpenstackProviderSpec).

HIVE-2308

(cherry picked from commit 28dac77cfc2d836956972ea9f2254de962ce63af)