Unstructured OpenstackProviderSpec
This is a manual backport of openshift#2196 to address the move of
OpenstackProviderSpec from cluster-api-provider-openstack to
openstack/api/machine.

We need to decode OpenstackProviderSpec to steal the osImage from the
master machines so we can use that value for new workers. Rather than
importing both schemata, trying to figure out which version we're
talking to, and decoding against the correct one, this commit unmarshals
the providerSpec as raw JSON into an Unstructured object and explicitly
paths into it map-wise.

On the creation side, we're told that the spoke cluster itself will
understand the new thing even if it's an old version (MAPI also
unmarshals the JSON raw). So it should be fine to continue using the
existing version of vendored installer code to generate MachineSets.

HIVE-2308