Nix and stop generating informers, listers
We haven't gone and confirmed this, but we think that
k8s.io/code-generator's `generate-groups.sh` when invoked with `"all"`
used to generate fewer things, and as they added things over the years,
we ended up generating them and checking them into our repository even
though we didn't need them and never used them.

`informers` and `listers` are two of these things.

This commit gets rid of the code those generators generated, and
switches our invocation of `generate-groups.sh` to explicitly list the
generators we intend to invoke, now omitting those two.

HIVE-2417