Vendor cpsm and gjson/sjson packages
A reworking of how we match up generated and remote MachineSets now
requires access to the availability zone (failure domain) of a
MachineSet using the
cluster-control-plane-machine-set-operator ProviderConfig interface.

Additionally, the yaml-patch package has been abandoned and for the
purposes of this solution, vendoring of gjson and sjson packages as
replacements is necessary.

Vendor all necessary packages.

HIVE-2254

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>