Delete MachineAutoscalers that would get MaxReplicas==0
MachineAutoscalers are not allowed to have MaxReplicas==0.

PR #2215 / bce2d472 partially fixed the scenario where a MachinePool's
autoscaling maxReplicas is less than the number of AZs by causing such
MAs not to be *created*. However, when reducing the MachinePool's
maxReplicas to below the number of AZs, we were still trying to update
*existing* MAs to have MaxReplicas==0. This commit adjusts the logic to
delete them instead.

HIVE-2415