WIP: Fix OpenStack MachinePool unit tests
The upstream installer MachineSet generator wants to make an actual API
call to the OpenStack cloud to determine "trunk support" (whatever that
may be). There's no good way to mock, fake, or work around this in unit
tests. This commit proposes an upstream change to facilitate a
workaround.

FOR COMMENT/REFERENCE ONLY. DO NOT MERGE.

Needed by HIVE-2476