Refactor OpenStack MachineSets trunk support discovery
A prior commit revendored installer to pull in
https://github.com/openshift/installer/pull/8187 and
https://github.com/openshift/installer/pull/8209 which decoupled the
OpenStack cloud call for trunk support discovery from the
`MachineSets()` generator.

This commit refactors our invocation of that generator accordingly, and
repairs our long-broken unit test framework around it by attaching a
`trunkSupportDiscoverer` func to the actuator. In the production code
path, this is set to the `CheckNetworkExtensionAvailability` function
publicized by the aforementioned installer PRs, whereas in the unit test
path it is set to a no-op shim. Future tests can make this shim return
any permutation of possible values to exercise different code paths.

Subsequent commits will enhance this unit test suite to reproduce the
issue in the referenced card.

Part of HIVE-2476