MachinePools/AWS: Match by subnet ID only
The MachinePool controller compares generated MachineSets with existing
remote MachineSets to figure out which to add, update, or delete.

We used to do this matching based on naming conventions, but those have
been changing upstream, so we recently (HIVE-2254 / #2179) changed to
matching by label + failure domain. The latter is extracted using code
from control-plane-machine-set.

Unfortunately, in AWS, the same failure domain can contain a subnet
represented in three different ways:
- ID (canonical)
- ARN (canonical; not used)
- Filters (ambiguous)

If the generated and remote MachineSets differ in either the way the
failure domain subnet is represented (e.g. ID vs Filters) *or* both are
using Filters that differ despite resolving to the same subnet, our
existing algorithm will fail to match them, and the reconcile will fail
quietly (remote MachineSets are not updated, local MachinePool status is
not changed).

With this commit, we disambiguate by ensuring both sides represent their
failure domain subnets by ID. If the remote MachineSet uses filters, we
update it (MAPI won't act on that change, but it should be functionally
a no-op) so we won't have to look it up again next time. *TODO:* We
should also figure out a way to cache the subnet IDs for the MachinePool
for the same reason -- perhaps by updating the MachinePool itself.

HIVE-2443