Refactor MachinePool unit tests to use Builder/Options
In preparation for expanding OpenStack-specific MachinePool unit tests,
refactor the helper funcs in the existing suite to use the MachinePool
"Builder" pattern with "Options". For example, we change this:

```go
machinePool: func() *hivev1.MachinePool {
        mp := testMachinePoolWithoutLabelsTaints()
        mp.Spec.Labels = make(map[string]string)
        mp.Spec.Labels["test-label"] = "test-value"
        return mp
}(),
```

to:

```go
machinePool: testMachinePoolWithoutLabelsTaints(
        testmp.WithLabels(map[string]string{
                "test-label": "test-value",
        }),
),
```

This makes the test code slightly tighter and more readable, and brings
the suite more in line with our usual (or at least, best practice)
patterns.

Facilitates HIVE-2476