Accommodate CAPI SecurityGroups from installer MachineSets
A previous commit revendored installer which picked up
https://github.com/openshift/installer/pull/8006 which adds two default
security group filters. We need to take this into account when the
`extra-worker-security-group` annotation is in play, as we need to add
that additional SG name and its VPC filter to *all* the SecurityGroups
in the worker MachineSets. This needs to be done both for day 0 (install
manager) and day 2 (MachinePool controller).

Carried along with HIVE-2476