MachinePools/OpenStack: UT for nil pointer exception fix
When MachinePool.Spec.Replicas is unset, as it is when we're doing
autoscaling, this used to trigger a nil pointer exception in the
vendored installer code in the MachineSets function, which used to
assume it would never be nil.

A prior commit revendored installer to pick up a fix whereby that
assumption is no longer relied upon.

Add unit testing to prove that it works.

HIVE-2476